Sua Supplier

+++ src/js/supplier/details/supplier-details.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory, useLocation } from "react-router-dom";
-
+import Swal from "sweetalert2";
 //css
 import "../supplier.css";
 
@@ -38,24 +38,46 @@ export default function SupplierDetails() {
 
     setSupplier(supplierStore);
   }
-  function onClickSave() {
-    const data = {
-      supplierId: supplier.id,
-      supplierName: supplier.supplierName,
-      description: supplier.description,
-      address: supplier.address,
-      salePersonName: supplier.salePersonName,
-      phoneNumber: supplier.phoneNumber,
-      email: supplier.email,
-    };
 
-    console.log(data);
-    dispatch(updateDetailsSupplierAction({ data: data, token: token }));
+  function onSaveClick(event) {
+    event.preventDefault();
+    const form = document.getElementById("productDetailsForm");
+    if (!form.checkValidity()) {
+      event.preventDefault();
+      event.stopPropagation;
+    } else {
+      const data = {
+        supplierId: supplier.id,
+        supplierName: supplier.supplierName,
+        description: supplier.description,
+        address: supplier.address,
+        salePersonName: supplier.salePersonName,
+        phoneNumber: supplier.phoneNumber,
+        email: supplier.email,
+      };
+      console.log(data);
+      dispatch(updateDetailsSupplierAction({ data: data, token: token }));
+    }
+    form.classList.add("was-validated");
   }
   function onClickDelete() {
-    dispatch(
-      deleteSupplierAction({ id: location.state.supplierId, token: token })
-    );
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: " #3085d6",
+      confirmButtonText: "Delete",
+      reverseButtons: true,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(
+          deleteSupplierAction({ id: location.state.supplierId, token: token })
+        );
+        // Swal.fire("Deleted!", "Your stocktake has been deleted.", "success");
+      }
+    });
   }
   function onChangeValue(event) {
     setSupplier({ ...supplier, [event.target.name]: event.target.value });
@@ -105,10 +127,10 @@ export default function SupplierDetails() {
         {
           isShow: true,
           title: "Save",
-          action: () => onClickSave(),
-          // action: () => clicktTest(),
-          class: "btn-primary",
-          // style: {},
+          class: " btn-primary",
+          action: (e) => onSaveClick(e),
+          type: "submit",
+          form: "productDetailsForm",
         },
       ];
     }
@@ -147,47 +169,6 @@ export default function SupplierDetails() {
 
   return (
     <div>
-      {/* todo: task heading */}
-      {/* <div className=" tab-fixed container-fluid  fixed-top">
-        <div className=" d-flex mb-3 justify-content-end mt-4 ">
-          <a className="me-2" onClick={goBackClick}>
-            <h3>Back</h3>
-          </a>
-          <h2 className="id-color fw-bold me-auto">Details</h2>
-          <div>
-            <button
-              className="btn btn-danger button-tab text-white button me-3"
-              onClick={onClickDelete}
-            >
-              Delete
-            </button>
-            {isDisabled ? (
-              <button
-                className="btn btn-warning button-tab text-white button me-3"
-                onClick={onClickEdit}
-              >
-                Edit
-              </button>
-            ) : (
-              <button
-                className="btn btn-secondary button-tab text-white button me-3"
-                onClick={onClickCancel}
-              >
-                Cancel
-              </button>
-            )}
-
-            <button
-              className="btn btn-primary button-tab button me-3"
-              disabled={isDisabled}
-              onClick={onClickSave}
-            >
-              Save
-            </button>
-          </div>
-        </div>
-      </div> */}
-
       <NavigationBar
         listButton={listButton}
         titleBar="Details"
@@ -195,62 +176,116 @@ export default function SupplierDetails() {
         status=""
       />
       {/* content */}
-      {returnData && (
-        <div className="wrapper space-top">
-          {/* content field 1  */}
-          <div className="wrapper-content shadow">
-            <div className="title-heading mt-2">
-              <span>Supplier Details</span>
-            </div>
-            <form>
-              {/* Supplier name  */}
-              <div className="mt-3">
-                <label for="supplierName" class="form-label">
-                  Supplier Name
-                </label>
-                <input
-                  type="text"
-                  class="form-control"
-                  name="supplierName"
-                  value={supplier.supplierName}
-                  disabled={isDisabled}
-                  readOnly={isDisabled}
-                  onChange={onChangeValue}
-                />
-              </div>
-              {/* Email &&  Phone No  */}
-              <div className="mt-3">
-                <div class="row g-3 align-items-center">
-                  <div class="col">
-                    <label for="inputEmail" class="col-form-label">
-                      Email
-                    </label>{" "}
-                    <input
-                      type="text"
-                      name="email"
-                      class="form-control"
-                      value={supplier.email}
-                      disabled={isDisabled}
-                      readOnly={isDisabled}
-                      onChange={onChangeValue}
-                    />
-                  </div>
-                  <div class="col">
-                    <label for="phone" class="col-form-label">
-                      Phone No.
-                    </label>{" "}
-                    <input
-                      type="tel"
-                      name="phoneNumber"
-                      class="form-control"
-                      value={supplier.phoneNumber}
-                      disabled={isDisabled}
-                      readOnly={isDisabled}
-                      onChange={onChangeValue}
-                    />
+      {/* {returnData && (
+        
+      )} */}
+      <div className="wrapper space-top">
+        <div className="card">
+          <div class="card-header">Supplier Information</div>
+          <ul class="list-group list-group-flush">
+            <li class="list-group-item">
+              <form
+                id="productDetailsForm"
+                class="row g-3 needs-validation "
+                noValidate
+              >
+                <div className="mt-3">
+                  <label for="supplierName" class="form-label">
+                    Name
+                  </label>
+                  <input
+                    type="text"
+                    class="form-control"
+                    name="supplierName"
+                    value={supplier.supplierName}
+                    disabled={isDisabled}
+                    readOnly={isDisabled}
+                    onChange={onChangeValue}
+                    required
+                    placeholder="etc. Nike Company"
+                  />
+                  <div class="invalid-feedback">Please input valid name</div>
+                </div>
+                <div className="mt-3">
+                  <label for="search" class="form-label">
+                    Seller Name
+                  </label>
+                  <input
+                    type="text"
+                    class="form-control"
+                    name="salePersonName"
+                    value={supplier.salePersonName}
+                    onChange={onChangeValue}
+                    disabled={isDisabled}
+                    readOnly={isDisabled}
+                    placeholder="Type seller name"
+                    required
+                  />
+                  <div className="invalid-feedback">
+                    PLease input valid seller name
                   </div>
                 </div>
-              </div>
+                {/* Email &&  Phone No  */}
+                <div className="mt-3">
+                  <div class="row g-3 align-items-center">
+                    <div class="col">
+                      <label for="inputEmail" class="col-form-label">
+                        Email
+                      </label>{" "}
+                      <input
+                        type="text"
+                        name="email"
+                        class="form-control"
+                        value={supplier.email}
+                        disabled={isDisabled}
+                        readOnly={isDisabled}
+                        onChange={onChangeValue}
+                        pattern="^[a-z][a-z0-9_\.]{5,32}@[a-z0-9]{2,}(\.[a-z0-9]{2,4}){1,2}$"
+                        required
+                        placeholder="clarice@gmail.com"
+                      />
+                      <div class="invalid-feedback">
+                        Please input valid email
+                      </div>
+                    </div>
+                    <div class="col">
+                      <label for="phone" class="col-form-label">
+                        Phone No.
+                      </label>{" "}
+                      <input
+                        type="tel"
+                        name="phoneNumber"
+                        class="form-control"
+                        value={supplier.phoneNumber}
+                        disabled={isDisabled}
+                        readOnly={isDisabled}
+                        onChange={onChangeValue}
+                        pattern="((09|03|07|08|05|028|024)+([0-9]{8})\b)"
+                        required
+                        placeholder="0903321332"
+                      />
+                      <div class="invalid-feedback">
+                        Please input valid phone number
+                      </div>
+                    </div>
+                  </div>
+                </div>
+                <div className="mb-3 mt-3">
+                  <label for="note" class="form-label">
+                    Address
+                  </label>
+                  <input
+                    type="text"
+                    name="address"
+                    class="form-control"
+                    value={supplier.address}
+                    disabled={isDisabled}
+                    readOnly={isDisabled}
+                    onChange={onChangeValue}
+                  />
+                  <div class="invalid-feedback">Please input address</div>
+                </div>
+              </form>
               {/* Note  */}
               <div class="mb-3 mt-3">
                 <label for="note" class="form-label">
@@ -266,26 +301,10 @@ export default function SupplierDetails() {
                   onChange={onChangeValue}
                 ></textarea>
               </div>
-            </form>
-          </div>
-
-          {/* content field 2  */}
-          <div className="wrapper-content mt-3 shadow">
-            <div className="title-heading mt-2">
-              <span>Address</span>
-            </div>
-            <input
-              type="text"
-              name="address"
-              class="form-control"
-              value={supplier.address}
-              disabled={isDisabled}
-              readOnly={isDisabled}
-              onChange={onChangeValue}
-            />
-          </div>
+            </li>
+          </ul>
         </div>
-      )}
+      </div>
     </div>
   );
 }
